fix: add receipt endpoint
